Move validateConfirmedExternalTransaction into its own file
