+++ packages/transaction-controller/src/utils.ts
@@ -3,7 +3,6 @@ import {
   isValidHexAddress,
 } from '@metamask/controller-utils';
 import type { Hex } from '@metamask/utils';
-import { ethErrors } from 'eth-rpc-errors';
 import { addHexPrefix, isHexString } from 'ethereumjs-util';
 import type { Transaction as NonceTrackerTransaction } from 'nonce-tracker/dist/NonceTracker';
 
@@ -11,8 +10,7 @@ import type {
   GasPriceValue,
   FeeMarketEIP1559Values,
 } from './TransactionController';
-import { TransactionStatus } from './types';
-import type { Transaction, TransactionMeta } from './types';
+import type { TransactionStatus, Transaction, TransactionMeta } from './types';
 
 export const ESTIMATE_GAS_ERROR = 'eth_estimateGas rpc method error';
 
@@ -232,52 +230,4 @@ export function transactionMatchesNetwork(
     return transaction.networkID === networkId;
   }
   return false;
-}
-
-/**
- * Validates the external provided transaction meta.
- *
- * @param transactionMeta - The transaction meta to validate.
- * @param confirmedTxs - The confirmed transactions in controller state.
- * @param pendingTxs - The submitted transactions in controller state.
- */
-export function validateConfirmedExternalTransaction(
-  transactionMeta?: TransactionMeta,
-  confirmedTxs?: TransactionMeta[],
-  pendingTxs?: TransactionMeta[],
-) {
-  if (!transactionMeta || !transactionMeta.transaction) {
-    throw ethErrors.rpc.invalidParams(
-      '"transactionMeta" or "transactionMeta.transaction" is missing',
-    );
-  }
-
-  if (transactionMeta.status !== TransactionStatus.confirmed) {
-    throw ethErrors.rpc.invalidParams(
-      'External transaction status should be "confirmed"',
-    );
-  }
-
-  const externalTxNonce = transactionMeta.transaction.nonce;
-  if (pendingTxs && pendingTxs.length > 0) {
-    const foundPendingTxByNonce = pendingTxs.find(
-      (tx) => tx.transaction?.nonce === externalTxNonce,
-    );
-    if (foundPendingTxByNonce) {
-      throw ethErrors.rpc.invalidParams(
-        'External transaction nonce should not be in pending txs',
-      );
-    }
-  }
-
-  if (confirmedTxs && confirmedTxs.length > 0) {
-    const foundConfirmedTxByNonce = confirmedTxs.find(
-      (tx) => tx.transaction?.nonce === externalTxNonce,
-    );
-    if (foundConfirmedTxByNonce) {
-      throw ethErrors.rpc.invalidParams(
-        'External transaction nonce should not be in confirmed txs',
-      );
-    }
-  }
 }
